+++ source/tridiagonal_matrix/tridiagonal_matrix_solver.h
@@ -5,11 +5,12 @@
 
 #include "tridiagonal_matrix.h"
 
-namespace trd {
+namespace trd
+{
 // Function to solve SoLE
-    std::vector<double> SolveTridiagonalMatrix(unsigned int order,
-                                               const TridiagonalMatrix &matrix,
-                                               const std::vector<double> &vector);
+    std::vector<double> SolveTridiagonalMatrix(size_t order,
+                                               const TridiagonalMatrix& matrix,
+                                               const std::vector<double>& vector);
 }
 
 #endif //SOLE_TRIDIAGONAL_MATRIX_SOLVER_H
